fixed integer issue

+++ db/migrate/20221226005902_create_members.rb
@@ -3,7 +3,7 @@ class CreateMembers < ActiveRecord::Migration[7.0]
     create_table :members do |t|
       t.string :name
       t.string :email
-      t.integer :phone
+      t.bigint :phone
 
       t.timestamps
     end
